Begin massive revamp of problem handling
The current problem handling system does not conform to the NCIP spec
very strongly. This is the start of handinling it using the correct
NCIP "Problem" definitions, as found in the NCIP 2.02 Protocol.
http://www.ncip.info/uploads/7/1/4/6/7146749/z39-83-1-2012_ncip.pdf

The old system is largely still in place for now. The new system accepts
an 'problems' template parameter that is an array of hashes with the keys
being the valid problem fields according to the current ncip schema.