Bug Fix - NCIP server cannot handle XML streams without line breaks
For some reason, if you remove all the line breaks from an XML message,
the NCIP server chokes. It looks like this may be an issue with the XML
parser we are using. Depending on line breaks, the NCIP message tag
may be the first or the second child node!

Resolved by checking the first node, then the second. I do not believe
there is a situation where the message tag would not be in one of the
first two child nodes.